requirements.txt
streamlit
pandas
plotly
langchain
langchain-groq
langchain-experimental
langchain-community
sqlalchemy
python-dotenv
groq
